# proyect2/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Wild Fantasy API",
    "description": "API documentation for managing characters, factions, places, species, items, and more in the Wild Fantasy universe.",
    "version": "1.0.0",
    "contact": {
      "name": "Joshua Burkhan"
    }
  },
  "host": "localhost:3000",
  "servers": [
    "http://localhost:3000"
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/newworld/characters/{id}": {
      "get": {
        "tags": [
          "Characters"
        ],
        "description": "Retrieve a character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Character ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Character data",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "age": {
                  "type": "string",
                  "example": "number"
                },
                "gender": {
                  "type": "string",
                  "example": "string"
                },
                "race": {
                  "type": "string",
                  "example": "string"
                },
                "nickname": {
                  "type": "string",
                  "example": "string"
                },
                "appearance": {
                  "type": "object",
                  "properties": {
                    "height": {
                      "type": "string",
                      "example": "number"
                    },
                    "weight": {
                      "type": "string",
                      "example": "number"
                    },
                    "eyeColor": {
                      "type": "string",
                      "example": "string"
                    },
                    "hairColor": {
                      "type": "string",
                      "example": "string"
                    },
                    "clothingStyle": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                },
                "personality": {
                  "type": "object",
                  "properties": {
                    "traits": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "strengths": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "weaknesses": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "quirks": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "history": {
                  "type": "object",
                  "properties": {
                    "birthplace": {
                      "type": "string",
                      "example": "string"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "year": {
                            "type": "string",
                            "example": "number"
                          },
                          "description": {
                            "type": "string",
                            "example": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "relationships": {
                  "type": "object",
                  "properties": {
                    "family": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "friends": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "enemies": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "allies": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "romance": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "abilities": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "string"
                      },
                      "elements": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "element": {
                              "type": "string",
                              "example": "string"
                            },
                            "orbs": {
                              "type": "string",
                              "example": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "coreRank": {
                  "type": "string",
                  "example": "string"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Character not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Character not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Server error"
                },
                "error": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Characters"
        ],
        "description": "Update an existing character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Character ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated character data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "age": {
                  "type": "number",
                  "example": 0
                },
                "gender": {
                  "type": "string",
                  "example": "string"
                },
                "race": {
                  "type": "string",
                  "example": "string"
                },
                "nickname": {
                  "type": "string",
                  "example": "string"
                },
                "appearance": {
                  "type": "object",
                  "properties": {
                    "height": {
                      "type": "number",
                      "example": 0
                    },
                    "weight": {
                      "type": "number",
                      "example": 0
                    },
                    "eyeColor": {
                      "type": "string",
                      "example": "string"
                    },
                    "hairColor": {
                      "type": "string",
                      "example": "string"
                    },
                    "clothingStyle": {
                      "type": "string",
                      "example": "string"
                    }
                  }
                },
                "personality": {
                  "type": "object",
                  "properties": {
                    "traits": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "strengths": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "weaknesses": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "quirks": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "history": {
                  "type": "object",
                  "properties": {
                    "birthplace": {
                      "type": "string",
                      "example": "string"
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "year": {
                            "type": "number",
                            "example": 0
                          },
                          "description": {
                            "type": "string",
                            "example": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "relationships": {
                  "type": "object",
                  "properties": {
                    "family": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "friends": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "enemies": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "allies": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    },
                    "romance": {
                      "type": "array",
                      "example": [
                        "string"
                      ],
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                },
                "abilities": {
                  "type": "array",
                  "example": [
                    "string"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "coreRank": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error updating character",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Error updating character"
                    },
                    "error": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Character not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Character not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Characters"
        ],
        "description": "Delete a character by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Character ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Character deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Character deleted successfully"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Character not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Character not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error deleting character",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error deleting character"
                },
                "error": {
                  "type": "object",
                  "properties": {}
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}